requirements.txt
gensim == 0.12.4
nltk == 3.2
pdfminer == 20140328
stop_words == 2015.2.23.1
Pillow == 3.2.0
numpy == 1.11.0
scipy == 0.17.1
scikit_learn == 0.17.1
